+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rules/Delegs.hs
@@ -20,12 +20,15 @@ module Cardano.Ledger.Shelley.Rules.Delegs (
   ShelleyDelegsPredFailure (..),
   ShelleyDelegsEvent (..),
   PredicateFailure,
-  delegsTransition,
+  validateZeroRewards,
+  validateDelegationRegistered,
+  drainWithdrawals,
 )
 where
 
 import Cardano.Ledger.Address (mkRwdAcnt)
 import Cardano.Ledger.BaseTypes (
+  Network,
   ShelleyBase,
   TxIx,
   invalidKey,
@@ -42,16 +45,22 @@ import Cardano.Ledger.Coin (Coin)
 import Cardano.Ledger.Core
 import Cardano.Ledger.Credential (Credential)
 import Cardano.Ledger.Keys (KeyHash, KeyRole (..))
+import Cardano.Ledger.Rules.ValidationMode (Test)
 import Cardano.Ledger.Shelley.Era (ShelleyDELEGS)
 import Cardano.Ledger.Shelley.LedgerState (
   AccountState,
   CertState (..),
+  DState (..),
   certDState,
-  dsUnified,
   psStakePoolParams,
   rewards,
  )
-import Cardano.Ledger.Shelley.Rules.Delpl (DelplEnv (..), ShelleyDELPL, ShelleyDelplEvent, ShelleyDelplPredFailure)
+import Cardano.Ledger.Shelley.Rules.Delpl (
+  DelplEnv (..),
+  ShelleyDELPL,
+  ShelleyDelplEvent,
+  ShelleyDelplPredFailure,
+ )
 import Cardano.Ledger.Shelley.TxBody (
   DCert (..),
   DelegCert (..),
@@ -75,8 +84,7 @@ import Control.State.Transition (
   judgmentContext,
   liftSTS,
   trans,
-  (?!),
-  (?!:),
+  validateTrans,
  )
 import Data.Map.Strict (Map)
 import qualified Data.Map.Strict as Map
@@ -87,6 +95,7 @@ import Data.Word (Word8)
 import GHC.Generics (Generic)
 import Lens.Micro ((^.))
 import NoThunks.Class (NoThunks (..))
+import Validation (failureUnless)
 
 data DelegsEnv era = DelegsEnv
   { delegsSlotNo :: !SlotNo
@@ -103,11 +112,14 @@ deriving stock instance
   Show (DelegsEnv era)
 
 data ShelleyDelegsPredFailure era
-  = DelegateeNotRegisteredDELEG
-      !(KeyHash 'StakePool (EraCrypto era)) -- target pool which is not registered
-  | WithdrawalsNotInRewardsDELEGS
-      !(Map (RewardAcnt (EraCrypto era)) Coin) -- withdrawals that are missing or do not withdrawal the entire amount
-  | DelplFailure (PredicateFailure (EraRule "DELPL" era)) -- Subtransition Failures
+  = -- | Target pool which is not registered
+    DelegateeNotRegisteredDELEG
+      !(KeyHash 'StakePool (EraCrypto era))
+  | -- | Withdrawals that are missing or do not withdrawal the entire amount
+    WithdrawalsNotInRewardsDELEGS
+      !(Map (RewardAcnt (EraCrypto era)) Coin)
+  | -- | Subtransition Failures
+    DelplFailure !(PredicateFailure (EraRule "DELPL" era))
   deriving (Generic)
 
 newtype ShelleyDelegsEvent era = DelplEvent (Event (EraRule "DELPL" era))
@@ -131,7 +143,7 @@ instance
   , Environment (EraRule "DELPL" era) ~ DelplEnv era
   , State (EraRule "DELPL" era) ~ CertState era
   , Signal (EraRule "DELPL" era) ~ DCert (EraCrypto era)
-  , PredicateFailure (EraRule "DELPL" era) ~ ShelleyDelplPredFailure era
+  , EraRule "DELEGS" era ~ ShelleyDELEGS era
   ) =>
   STS (ShelleyDELEGS era)
   where
@@ -201,74 +213,82 @@ delegsTransition ::
   , Environment (EraRule "DELPL" era) ~ DelplEnv era
   , State (EraRule "DELPL" era) ~ CertState era
   , Signal (EraRule "DELPL" era) ~ DCert (EraCrypto era)
-  , PredicateFailure (EraRule "DELPL" era) ~ ShelleyDelplPredFailure era
+  , EraRule "DELEGS" era ~ ShelleyDELEGS era
   ) =>
   TransitionRule (ShelleyDELEGS era)
 delegsTransition = do
-  TRC (env@(DelegsEnv slot txIx pp tx acnt), dpstate, certificates) <- judgmentContext
+  TRC (env@(DelegsEnv slot txIx pp tx acnt), certState, certificates) <- judgmentContext
   network <- liftSTS $ asks networkId
 
   case certificates of
     Empty -> do
-      let ds = certDState dpstate
-          withdrawals_ = unWithdrawals (tx ^. bodyTxL . withdrawalsTxBodyL)
-          rewards' = rewards ds
-      isSubmapOf withdrawals_ rewards' -- withdrawals_ ⊆ rewards
-        ?! WithdrawalsNotInRewardsDELEGS
-          ( Map.differenceWith
-              (\x y -> if x /= y then Just x else Nothing)
-              withdrawals_
-              (Map.mapKeys (mkRwdAcnt network) (UM.rewView (dsUnified ds)))
-          )
-
-      let drainedRewardAccounts :: Map.Map (Credential 'Staking (EraCrypto era)) UM.RDPair
-          drainedRewardAccounts =
-            Map.foldrWithKey
-              ( \(RewardAcnt _ cred) _coin ->
-                  Map.insert cred (UM.RDPair (UM.CompactCoin 0) (UM.CompactCoin 0))
-                  -- Note that the deposit (CompactCoin 0) will be ignored.
-              )
-              Map.empty
-              withdrawals_
-          unified' = rewards' UM.⨃ drainedRewardAccounts
-      pure $ dpstate {certDState = ds {dsUnified = unified'}}
+      let dState = certDState certState
+          withdrawals = tx ^. bodyTxL . withdrawalsTxBodyL
+      validateTrans WithdrawalsNotInRewardsDELEGS $
+        validateZeroRewards dState withdrawals network
+      pure $ certState {certDState = drainWithdrawals dState withdrawals}
     gamma :|> c -> do
-      dpstate' <-
-        trans @(ShelleyDELEGS era) $ TRC (env, dpstate, gamma)
-
-      let isDelegationRegistered = case c of
-            DCertDeleg (Delegate deleg) ->
-              let stPools_ = psStakePoolParams $ certPState dpstate'
-                  targetPool = dDelegatee deleg
-               in if eval (targetPool ∈ dom stPools_)
-                    then Right ()
-                    else Left $ DelegateeNotRegisteredDELEG targetPool
-            _ -> Right ()
-      isDelegationRegistered ?!: id
-
+      certState' <-
+        trans @(ShelleyDELEGS era) $ TRC (env, certState, gamma)
+      validateTrans DelegateeNotRegisteredDELEG $
+        validateDelegationRegistered certState' c
       -- It is impossible to have 65535 number of certificates in a
       -- transaction, therefore partial function is justified.
       let ptr = Ptr slot txIx (mkCertIxPartial $ toInteger $ length gamma)
       trans @(EraRule "DELPL" era) $
-        TRC (DelplEnv slot ptr pp acnt, dpstate', c)
+        TRC (DelplEnv slot ptr pp acnt, certState', c)
+
+validateDelegationRegistered ::
+  CertState era ->
+  DCert (EraCrypto era) ->
+  Test (KeyHash 'StakePool (EraCrypto era))
+validateDelegationRegistered certState = \case
+  DCertDeleg (Delegate deleg) ->
+    let stPools = psStakePoolParams $ certPState certState
+        targetPool = dDelegatee deleg
+     in failureUnless (eval (targetPool ∈ dom stPools)) targetPool
+  _ -> pure ()
+
+-- @withdrawals_@ is small and @rewards@ big, better to transform the former
+-- than the latter into the right shape so we can call 'Map.isSubmapOf'.
+isSubmapOfUM ::
+  forall era.
+  Map (RewardAcnt (EraCrypto era)) Coin ->
+  View (EraCrypto era) (Credential 'Staking (EraCrypto era)) UM.RDPair ->
+  Bool
+isSubmapOfUM ws (RewardDeposits (UMap tripmap _)) = Map.isSubmapOfBy f withdrawalMap tripmap
   where
-    -- @withdrawals_@ is small and @rewards@ big, better to transform the former
-    -- than the latter into the right shape so we can call 'Map.isSubmapOf'.
-    isSubmapOf ::
-      Map (RewardAcnt (EraCrypto era)) Coin ->
-      View (EraCrypto era) (Credential 'Staking c) UM.RDPair ->
-      Bool
-    isSubmapOf withdrawals_ (RewardDeposits (UMap tripmap _)) = Map.isSubmapOfBy f withdrawalMap tripmap
-      where
-        withdrawalMap :: Map.Map (Credential 'Staking (EraCrypto era)) Coin
-        withdrawalMap =
-          Map.fromList
-            [ (cred, coin)
-            | (RewardAcnt _ cred, coin) <- Map.toList withdrawals_
-            ]
-        f :: Coin -> Trip (EraCrypto era) -> Bool
-        f coin1 (Triple (SJust (UM.RDPair coin2 _)) _ _) = coin1 == fromCompact coin2
-        f _ _ = False
+    withdrawalMap :: Map.Map (Credential 'Staking (EraCrypto era)) Coin
+    withdrawalMap = Map.mapKeys (\(RewardAcnt _ cred) -> cred) ws
+    f :: Coin -> Trip (EraCrypto era) -> Bool
+    f coin1 (Triple (SJust (UM.RDPair coin2 _)) _ _) = coin1 == fromCompact coin2
+    f _ _ = False
+
+drainWithdrawals :: DState era -> Withdrawals (EraCrypto era) -> DState era
+drainWithdrawals dState (Withdrawals wdrls) =
+  dState {dsUnified = rewards dState UM.⨃ drainedRewardAccounts}
+  where
+    drainedRewardAccounts =
+      Map.foldrWithKey
+        ( \(RewardAcnt _ cred) _coin ->
+            Map.insert cred (UM.RDPair (UM.CompactCoin 0) (UM.CompactCoin 0))
+            -- Note that the deposit (CompactCoin 0) will be ignored.
+        )
+        Map.empty
+        wdrls
+
+validateZeroRewards ::
+  forall era.
+  DState era ->
+  Withdrawals (EraCrypto era) ->
+  Network ->
+  Test (Map (RewardAcnt (EraCrypto era)) Coin)
+validateZeroRewards dState (Withdrawals wdrls) network = do
+  failureUnless (isSubmapOfUM @era wdrls (rewards dState)) $ -- withdrawals_ ⊆ rewards
+    Map.differenceWith
+      (\x y -> if x /= y then Just x else Nothing)
+      wdrls
+      (Map.mapKeys (mkRwdAcnt network) (UM.rewView (dsUnified dState)))
 
 instance
   ( Era era